@@ -17,15 +17,14 @@ android {
             storePassword '53#falls'
         }
     }
-    compileSdkVersion 28
-    //buildToolsVersion '28.0.3'
+    compileSdkVersion 30
     defaultConfig {
         applicationId "com.servir.invasivespecies"
-        minSdkVersion 16
-        targetSdkVersion 28
-        versionCode 10
-        versionName "ISMS_10_05082019"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        minSdkVersion 24
+        targetSdkVersion 30
+        versionCode 11
+        versionName "ISMS_11_19012022"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
     buildTypes {
         release {
@@ -33,28 +32,32 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
             signingConfig signingConfigs.ISMS_RELEASE
         }
-
+        useLibrary 'org.apache.http.legacy'
+        productFlavors {
+        }
         debug {
             signingConfig signingConfigs.ISMS_DEBUG
-            //debuggable true
         }
     }
-    useLibrary 'org.apache.http.legacy'
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
 }
 
 dependencies {
-    implementation fileTree(include: ['*.jar'], dir: 'libs')
-    androidTestImplementation('com.android.support.test.espresso:espresso-core:3.0.2', {
-        exclude group: 'com.android.support', module: 'support-annotations'
-    })
-    //noinspection GradleCompatible
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    //implementation 'com.android.support:support-media-compat:28.0.0'
-    implementation 'com.google.android.gms:play-services-location:16.0.0'
-    implementation 'com.google.android.gms:play-services-maps:16.1.0'
-    implementation 'org.apache.httpcomponents:httpcomponents-client:4.5.8'
-    implementation 'org.apache.httpcomponents:httpcomponents-core:4.4.11'
-    testImplementation 'junit:junit:4.13-beta-2'
-    implementation 'com.google.maps.android:android-maps-utils:0.5'
-    implementation 'com.google.firebase:firebase-appindexing:17.1.0'
+    implementation 'androidx.appcompat:appcompat:1.4.1'
+    implementation 'com.google.android.material:material:1.5.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.3'
+    implementation 'org.apache.httpcomponents:httpcomponents-client:4.5.13'
+    implementation 'org.apache.httpcomponents:httpcomponents-core:4.4.15'
+    implementation 'com.google.android.gms:play-services-maps:18.0.2'
+    implementation 'com.google.android.gms:play-services-location:19.0.1'
+    implementation 'com.google.maps.android:android-maps-utils:2.3.0'
+    implementation 'com.android.volley:volley:1.2.1'
+    implementation 'com.google.maps.android:android-maps-utils:2.3.0'
+    implementation 'com.google.firebase:firebase-appindexing:20.0.0'
+    testImplementation 'junit:junit:4.13.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
 }
